% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figures.R
\name{create_figures}
\alias{create_figures}
\title{Create figures}
\usage{
create_figures_models(fig = 1)
}
\description{
Creates a series of figures to aide the interpretation of model outputs.
The figures are model dependent and should only be used if the models have
converged correctly.
}
\details{
Figures include:
\enumerate{
\item{Multiplicative growth rates}
\item{Interaction coefficients}
\item{Fitness differences}
\item{Niche overlap}
\item{Competitive abilities}
\item{Posterior predictions}
}
}
