% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data}
\alias{format_data}
\title{Format data for analysis}
\usage{
format_data(model, n_sp = 4, ts = 12, ...)
}
\arguments{
\item{n_sp}{Treat residents as separate species (n=6), or single community (n=4).}

\item{ts}{Number of timesteps for iteration method.}
}
\description{
Takes the preloaded Pinnacle vegetation dataset and prepares it for
analysis in Stan. The Stan models are written such that the are able
to take largely the same data structure, with the exception of M0.
}
