% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models.R
\name{run_models}
\alias{run_models}
\title{Run models}
\usage{
run_model(model = "m1")
}
\arguments{
\item{n_sp}{Treat residents as separate species (n=6), or single community (n=4).}

\item{ts}{Number of timesteps for iteration method.}

\item{path}{Directory to save model outputs, defaults to /models.}

\item{check}{auto-run model checks, boolean (default = T).}

\item{model}{BV (~2hrs).}
}
\description{
Models are coded in Stan and can be found in the \code{/models} subdirectory.
Model output is saved to .Rdata files for later inspection. These are
are not distributed due to their size. Previously saved analysis can be
inspected using \code{check_models()} or loaded with \code{load_models()}.
}
\details{
Models include:
\enumerate{
\item{BV: Beverton-Holt biomass model}
}
}
