% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data}
\alias{format_data}
\title{Format data for analysis}
\usage{
format_data(...)
}
\value{
\itemize{
\item N    - the number of individuals
\item N_jk - the number of observed interactions
\item S    - the number of species
\item A    - the number of distinct interactions
\item P    - the number of blocks
\item y    - aboveground biomass measurements
\item jl   - IDs for species j in fertility treatment l
\item jkl  - IDs for interaction between species k on species j in treatment l
\item jp   - IDs for random effect of block p on species j
\item n_i  - number of competitors
}
}
\description{
Loads and prepares the experimental data for analysis in Stan.
}
\examples{
# load data
data_list <- format_data()
str(data_list)

}
